src/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-multipart>=0.0.6
aiofiles>=23.2.1
boto3>=1.28.63
httpx>=0.25.0
scrapy>=2.11.0
pytest>=7.4.2
python-dotenv>=1.0.0
logging>=0.4.9.6
asyncpg>=0.28.0 